+++ models.py
@@ -3,7 +3,8 @@ author: Sanidhya Mangal
 github: sanidhyamangal
 """
 import os  # for os related ops
-from functools import partial  # for creation of partial functions
+from functools import partial, reduce
+from typing import Tuple  # for creation of partial functions
 
 import tensorflow as tf  # for deep learning based options
 
@@ -12,7 +13,7 @@ class GenerativeModel(tf.keras.Model):
     """
     Generative model for conditional gans
     """
-    def __init__(self, *args, **kwargs):
+    def __init__(self, num_class:int=10, *args, **kwargs):
         super(GenerativeModel, self).__init__(*args, **kwargs)
 
         # create partial methods for conv model
@@ -24,52 +25,49 @@ class GenerativeModel(tf.keras.Model):
         BN = partial(tf.keras.layers.BatchNormalization)
         LRU = partial(tf.keras.layers.LeakyReLU)
 
-        # create a genrative model
-        self.dense1 = tf.keras.layers.Dense(7 * 7 * 128, use_bias=False)
-        self.conditional_dense = tf.keras.layers.Dense(units=7 * 7 * 128)
+        # embedding layers and conditional model
+        self.conditional = tf.keras.models.Sequential([
+            tf.keras.layers.Embedding(num_class, 50),
+            tf.keras.layers.Dense(7*7),
+            tf.keras.layers.Reshape((7,7,1))
+        ])
 
-        self.bn1 = BN()
-        self.act1 = LRU()
+        self.pre_gen = tf.keras.models.Sequential([
+            tf.keras.layers.Dense(7*7*128),
+            tf.keras.layers.LeakyReLU(),
+            tf.keras.layers.Reshape((7,7,128))
+        ])
 
-        self.reshape = tf.keras.layers.Reshape((7, 7, 128))
+        self.gen = tf.keras.models.Sequential([
+            Conv2DT(filters=128, strides=(1,1)),
+            BN(),
+            LRU(),
+            Conv2DT(filters=64),
+            BN(),
+            LRU(),
+            Conv2DT(filters=1),
+            BN(),
+            LRU()
+        ])
 
-        self.deconv1 = Conv2DT(filters=128)
-        self.bn2 = BN()
-        self.act2 = LRU()
 
-        self.deconv2 = Conv2DT(filters=1)
-        self.bn3 = BN()
-        self.act3 = LRU()
 
     def call(self, inputs, labels, *args, **kwargs):
         training = kwargs.pop('training', True)
-
-        # combine labels and inputs
-        x = tf.concat(
-            [inputs,
-             self.conditional_dense(tf.one_hot(labels, depth=10))],
-            axis=1)
-        x = self.dense1(inputs)
-
-        x = self.act1(self.bn1(x, training=training))
-        x = self.reshape(x)
-
-        # first deconv and upscaling part
-        x = self.deconv1(x)
-        x = self.act2(self.bn2(training=training))
-
-        # second deconv and upscaling layer
-        x = self.deconv2(x)
-        x = self.act3(self.bn3(x, training=training))
+        conditional_vec = self.conditional(labels)
+        x = self.pre_gen(inputs)
+        x = tf.concat([x, conditional_vec], axis=-1)
+        x = self.gen(x, training=training)
 
         return x
+        
 
 
 class DiscrimintaiveModel(tf.keras.models.Model):
     """
     Class for perfroming discriminative models for the conditional lsgan
     """
-    def __init__(self, *args, **kwargs):
+    def __init__(self, input_shape:Tuple[int]=(28,28,1),num_classes:int=10,*args, **kwargs):
         super(DiscrimintaiveModel, self).__init__(*args, **kwargs)
 
         # create partial methods for model
@@ -79,41 +77,33 @@ class DiscrimintaiveModel(tf.keras.models.Model):
                          strides=(2, 2),
                          use_bias=False)
         FC = partial(tf.keras.layers.Dense)
-        BN = partial(tf.keras.layers.BatchNormalization)
-        self.conditional_dense = FC(units=256)
-        self.conv1 = Conv2D(filters=256)
+        Relu = partial(tf.keras.layers.ReLU)
+        self.conditional = tf.keras.models.Sequential([
+            tf.keras.layers.Embedding(num_classes, 50),
+            tf.keras.layers.Dense(reduce(lambda a,b: a*b, input_shape)),
+            tf.keras.layers.Reshape(input_shape)
+        ])
 
-        self.conv2 = Conv2D(filters=320)
-        self.bn1 = BN()
-
-        self.fc1 = FC(units=1024)
-        self.bn2 = BN()
-        self.fc2 = FC(units=1)
+        self.conv = tf.keras.models.Sequential([
+            Conv2D(filters=32),
+            Relu(),
+            tf.keras.layers.Dropout(rate=0.3),
+            Conv2D(filters=64),
+            Relu(),
+            tf.keras.layers.Dropout(rate=0.3),
+            Conv2D(filters=64),
+            Relu(),
+            tf.keras.layers.Dropout(rate=0.3),
+            tf.keras.layers.Flatten(),
+            FC(1)
+        ])
 
     def call(self, inputs, labels, *args, **kwargs):
         training = kwargs.pop('training', True)
+        
+        conditional_vector = self.conditional(labels)
 
-        conditional_vector = self.conditional_dense(
-            tf.one_hot(labels, depth=10))
-        conditional_vector_conv = tf.reshape(conditional_vector,
-                                             [-1, 1, 1, 256])
-        x = self.conv1(inputs)
-        x = self.conv_concat(x, conditional_vector_conv)
+        x = tf.concat([inputs, conditional_vector], axis=-1)
+        x = self.conv(x)
 
-        # conv layer ops 2
-        x = self.bn1(self.conv2(x), training=training)
-        x = tf.reshape(x, [x.shape[0], -1])
-
-        x = tf.concat([x, conditional_vector], axis=1)
-
-        x = self.bn2(self.fc1(x), training=training)
-
-        x = self.fc2(x)
         return x
-
-    @tf.function
-    def conv_concat(self, x: tf.Tensor, y: tf.Tensor):
-        x_shape = x.shape
-        y_shape = y.shape
-
-        return tf.concat([x, y * tf.ones(shape=x_shape)], axis=3)