+++ changelog.md
@@ -5,3 +5,6 @@
 - fixed issues with generative models
 - added discriminator model for conditional gans
 - added data handler pipeline for loading data
+
+### 2021-02-06
+- fixed: cgan architecture for model